structure: test/application, application/port 도메인별로 패키징
