+++ src/main/java/com/ddudu/application/service/goal/DeleteGoalService.java
@@ -3,7 +3,7 @@ package com.ddudu.application.service.goal;
 import com.ddudu.application.annotation.UseCase;
 import com.ddudu.application.domain.goal.domain.Goal;
 import com.ddudu.application.domain.goal.exception.GoalErrorCode;
-import com.ddudu.application.port.in.DeleteGoalUseCase;
+import com.ddudu.application.port.in.goal.DeleteGoalUseCase;
 import com.ddudu.application.port.out.goal.DeleteGoalPort;
 import com.ddudu.application.port.out.goal.GoalLoaderPort;
 import java.util.MissingResourceException;
