+++ src/main/java/com/ddudu/presentation/api/controller/GoalController.java
@@ -7,9 +7,9 @@ import com.ddudu.application.domain.goal.dto.response.GoalIdResponse;
 import com.ddudu.application.domain.goal.dto.response.GoalResponse;
 import com.ddudu.application.domain.goal.dto.response.GoalSummaryResponse;
 import com.ddudu.application.domain.goal.exception.GoalErrorCode;
-import com.ddudu.application.port.in.DeleteGoalUseCase;
 import com.ddudu.application.port.in.goal.ChangeGoalStatusUseCase;
 import com.ddudu.application.port.in.goal.CreateGoalUseCase;
+import com.ddudu.application.port.in.goal.DeleteGoalUseCase;
 import com.ddudu.application.port.in.goal.RetrieveAllGoalsUseCase;
 import com.ddudu.application.port.in.goal.RetrieveGoalUseCase;
 import com.ddudu.application.port.in.goal.UpdateGoalUseCase;
