+++ src/test/java/com/ddudu/application/service/goal/ChangeGoalStatusServiceTest.java
@@ -1,27 +1,22 @@
-package com.ddudu.application.service;
+package com.ddudu.application.service.goal;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 import com.ddudu.application.domain.goal.domain.Goal;
+import com.ddudu.application.domain.goal.domain.enums.GoalStatus;
+import com.ddudu.application.domain.goal.dto.request.ChangeGoalStatusRequest;
 import com.ddudu.application.domain.goal.exception.GoalErrorCode;
 import com.ddudu.application.domain.user.domain.User;
 import com.ddudu.application.port.out.auth.SignUpPort;
-import com.ddudu.application.port.out.goal.DeleteGoalPort;
 import com.ddudu.application.port.out.goal.GoalLoaderPort;
 import com.ddudu.application.port.out.goal.SaveGoalPort;
 import com.ddudu.application.port.out.user.UserLoaderPort;
-import com.ddudu.application.service.goal.DeleteGoalService;
 import com.ddudu.fixture.BaseFixture;
-import com.ddudu.fixture.DduduFixture;
 import com.ddudu.fixture.GoalFixture;
 import com.ddudu.fixture.UserFixture;
-import com.ddudu.old.todo.domain.Todo;
-import com.ddudu.old.todo.domain.TodoRepository;
-import jakarta.persistence.EntityManager;
 import jakarta.transaction.Transactional;
 import java.util.MissingResourceException;
-import java.util.Optional;
 import org.assertj.core.api.ThrowableAssert.ThrowingCallable;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -33,10 +28,10 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 @Transactional
 @DisplayNameGeneration(ReplaceUnderscores.class)
-class DeleteGoalServiceTest {
+class ChangeGoalStatusServiceTest {
 
   @Autowired
-  DeleteGoalService deleteGoalService;
+  ChangeGoalStatusService changeGoalStatusService;
 
   @Autowired
   UserLoaderPort userLoaderPort;
@@ -50,74 +45,59 @@ class DeleteGoalServiceTest {
   @Autowired
   SaveGoalPort saveGoalPort;
 
-  @Autowired
-  DeleteGoalPort deleteGoalPort;
-
-  @Autowired
-  TodoRepository todoRepository;
-
-  @Autowired
-  EntityManager entityManager;
-
   Long userId;
   Goal goal;
+  ChangeGoalStatusRequest request;
+  GoalStatus newStatus;
 
   @BeforeEach
   void setUp() {
     User user = createAndSaveUser();
     userId = user.getId();
     goal = createAndSaveGoal(user);
+    newStatus = GoalFixture.getRandomGoalStatus();
+    request = new ChangeGoalStatusRequest(newStatus.name());
   }
 
   @Test
-  void 목표를_삭제_할_수_있다() {
+  void 목표_상태를_변경할_수_있다() {
     // when
-    deleteGoalService.delete(userId, goal.getId());
+    changeGoalStatusService.changeStatus(userId, goal.getId(), request);
 
     // then
-    Optional<Goal> foundAfterDeleted = goalLoaderPort.findById(goal.getId());
-    assertThat(foundAfterDeleted).isEmpty();
+    Goal actual = goalLoaderPort.findById(goal.getId())
+        .get();
+    assertThat(actual.getStatus()).isEqualTo(newStatus);
   }
 
   @Test
-  void 목표_삭제_시_해당_목표의_뚜두도_삭제된다() {
-    //given
-    Todo todo = DduduFixture.createRandomDduduWithGoal(goal);
-    todo = todoRepository.save(todo);
-
-    //when
-    deleteGoalService.delete(userId, goal.getId());
-    flushAndClear();
-
-    //then
-    assertThat(todoRepository.findById(todo.getId())).isEmpty();
-  }
-
-  @Test
-  void 목표가_존재하지_않는_경우_예외가_발생한다() {
+  void 유효하지_않은_ID인_경우_수정에_실패한다() {
     // given
     Long invalidId = BaseFixture.getRandomId();
 
     // when
-    ThrowingCallable delete = () -> deleteGoalService.delete(userId, invalidId);
+    ThrowingCallable update = () -> changeGoalStatusService.changeStatus(
+        userId, invalidId, request);
 
     // then
-    assertThatExceptionOfType(MissingResourceException.class).isThrownBy(delete)
+    assertThatExceptionOfType(MissingResourceException.class).isThrownBy(update)
         .withMessage(GoalErrorCode.ID_NOT_EXISTING.getCodeName());
   }
 
   @Test
-  void 로그인_사용자가_권한이_없는_경우_삭제에_실패한다() {
+  void 로그인_사용자가_권한이_없는_경우_수정에_실패한다() {
     // given
     User anotherUser = createAndSaveUser();
 
     // when
-    ThrowingCallable delete = () -> deleteGoalService.delete(anotherUser.getId(), goal.getId());
+    ThrowingCallable update = () -> changeGoalStatusService.changeStatus(
+        anotherUser.getId(), goal.getId(), request);
 
     // then
-    assertThatExceptionOfType(SecurityException.class).isThrownBy(delete)
+    assertThatExceptionOfType(SecurityException.class).isThrownBy(update)
         .withMessage(GoalErrorCode.INVALID_AUTHORITY.getCodeName());
   }
+
 
   private User createAndSaveUser() {
     User user = UserFixture.createRandomUserWithId();
@@ -127,11 +107,6 @@ class DeleteGoalServiceTest {
   private Goal createAndSaveGoal(User user) {
     Goal goal = GoalFixture.createRandomGoalWithUser(user);
     return saveGoalPort.save(goal);
-  }
-
-  private void flushAndClear() {
-    entityManager.flush();
-    entityManager.clear();
   }
 
 }
