+++ src/test/java/com/ddudu/application/service/goal/RetrieveAllGoalsServiceTest.java
@@ -1,22 +1,24 @@
-package com.ddudu.application.service;
+package com.ddudu.application.service.goal;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 import com.ddudu.application.domain.goal.domain.Goal;
-import com.ddudu.application.domain.goal.dto.response.GoalResponse;
+import com.ddudu.application.domain.goal.dto.response.GoalSummaryResponse;
 import com.ddudu.application.domain.goal.exception.GoalErrorCode;
 import com.ddudu.application.domain.user.domain.User;
 import com.ddudu.application.port.out.auth.SignUpPort;
 import com.ddudu.application.port.out.goal.GoalLoaderPort;
 import com.ddudu.application.port.out.goal.SaveGoalPort;
 import com.ddudu.application.port.out.user.UserLoaderPort;
-import com.ddudu.application.service.goal.RetrieveGoalService;
 import com.ddudu.fixture.BaseFixture;
 import com.ddudu.fixture.GoalFixture;
 import com.ddudu.fixture.UserFixture;
 import jakarta.transaction.Transactional;
+import java.util.Comparator;
+import java.util.List;
 import java.util.MissingResourceException;
+import java.util.stream.IntStream;
 import org.assertj.core.api.ThrowableAssert.ThrowingCallable;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayNameGeneration;
@@ -28,10 +30,10 @@ import org.springframework.boot.test.context.SpringBootTest;
 @SpringBootTest
 @Transactional
 @DisplayNameGeneration(ReplaceUnderscores.class)
-class RetrieveGoalServiceTest {
+class RetrieveAllGoalsServiceTest {
 
   @Autowired
-  RetrieveGoalService retrieveGoalService;
+  RetrieveAllGoalsService retrieveAllGoalsService;
 
   @Autowired
   UserLoaderPort userLoaderPort;
@@ -46,50 +48,46 @@ class RetrieveGoalServiceTest {
   SaveGoalPort saveGoalPort;
 
   Long userId;
-  Goal goal;
+
+  List<Goal> goals;
 
   @BeforeEach
   void setUp() {
     User user = createAndSaveUser();
     userId = user.getId();
-    goal = createAndSaveGoal(user);
+    goals = createAndSaveGoals(user);
   }
 
   @Test
-  void ID를_통해_목표를_조회_할_수_있다() {
+  void 사용자의_전체_목표를_조회_할_수_있다() {
     // when
-    GoalResponse actual = retrieveGoalService.getById(userId, goal.getId());
+    List<GoalSummaryResponse> actual = retrieveAllGoalsService.findAllByUser(userId);
 
     // then
-    assertThat(actual).extracting("id", "name", "status", "color", "privacyType")
-        .containsExactly(
-            goal.getId(), goal.getName(), goal.getStatus(), goal.getColor(), goal.getPrivacyType());
+    assertThat(actual.size()).isEqualTo(goals.size());
+    for (int i = 0; i < goals.size(); i++) {
+      assertThat(actual.get(i))
+          .extracting("id", "name", "status", "color")
+          .containsExactly(goals.get(i)
+              .getId(), goals.get(i)
+              .getName(), goals.get(i)
+              .getStatus(), goals.get(i)
+              .getColor());
+    }
   }
 
   @Test
-  void 유효하지_않은_ID인_경우_조회에_실패한다() {
+  void 사용자가_존재하지_않는_경우_조회에_실패한다() {
     // given
-    Long invalidId = BaseFixture.getRandomId();
+    Long invalidLoginId = BaseFixture.getRandomId();
 
     // when
-    ThrowingCallable getById = () -> retrieveGoalService.getById(userId, invalidId);
+    ThrowingCallable findAllByUser = () -> retrieveAllGoalsService.findAllByUser(
+        invalidLoginId);
 
     // then
-    assertThatExceptionOfType(MissingResourceException.class).isThrownBy(getById)
-        .withMessage(GoalErrorCode.ID_NOT_EXISTING.getCodeName());
-  }
-
-  @Test
-  void 로그인_사용자가_목표의_주인이_아닌_경우_조회에_실패한다() {
-    // given
-    User anotherUser = createAndSaveUser();
-
-    // when
-    ThrowingCallable getById = () -> retrieveGoalService.getById(anotherUser.getId(), goal.getId());
-
-    // then
-    assertThatExceptionOfType(SecurityException.class).isThrownBy(getById)
-        .withMessage(GoalErrorCode.INVALID_AUTHORITY.getCodeName());
+    assertThatExceptionOfType(MissingResourceException.class).isThrownBy(findAllByUser)
+        .withMessage(GoalErrorCode.USER_NOT_EXISTING.getCodeName());
   }
 
   private User createAndSaveUser() {
@@ -97,9 +95,13 @@ class RetrieveGoalServiceTest {
     return signUpPort.save(user);
   }
 
-  private Goal createAndSaveGoal(User user) {
-    Goal goal = GoalFixture.createRandomGoalWithUser(user);
-    return saveGoalPort.save(goal);
+  private List<Goal> createAndSaveGoals(User user) {
+    return IntStream.range(0, 3)
+        .mapToObj(i -> GoalFixture.createRandomGoalWithUser(user))
+        .map(saveGoalPort::save)
+        .sorted(Comparator.comparingLong(Goal::getId)
+            .reversed())
+        .toList();
   }
 
 }
